+++ src/identity_map/identity_map/test.rs
@@ -1,30 +1,4 @@
 // Copyright 2025 Gabriel Bjørnager Jensen.
-//
-// Permission is hereby granted, free of charge, to
-// any person obtaining a copy of this software and
-// associated documentation files (the "Software"),
-// to deal in the Software without restriction, in-
-// cluding without limitation the rights to use,
-// copy, modify, merge, publish, distribute, subli-
-// cense, and/or sell copies of the Software, and
-// to permit persons to whom the Software is fur-
-// nished to do so, subject to the following condi-
-// tions:
-//
-// The above copyright notice and this permission
-// notice shall be included in all copies or sub-
-// stantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WAR-
-// RANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUD-
-// ING BUT NOT LIMITED TO THE WARRANTIES OF MER-
-// CHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE,
-// AND NONINFRINGEMENT. IN NO EVENT SHALL THE AU-
-// THORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-// CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
-// ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
-// OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 use crate::IdentityMap;
 
@@ -32,7 +6,7 @@ use alloc::vec::Vec;
 use bincode::{deserialize_from, serialize_into};
 use core::sync::atomic::{AtomicU8, Ordering};
 
-#[allow(clippy::len_zero)]
+#[expect(clippy::len_zero)]
 #[test]
 fn test_identity_map() {
 	assert_eq!(size_of::<IdentityMap<usize, u8>>(), size_of::<Option<IdentityMap<(),()>>>());
@@ -293,11 +267,25 @@ fn test_identity_map_iter() {
 }
 
 #[test]
+fn test_identity_map_pop() {
+	let mut data: IdentityMap<u8, u8> = (u8::MIN..=u8::MAX).map(|k| (k, u8::MAX - k)).collect();
+
+	assert_eq!(data.pop_first(), Some((0x00, 0xFF)));
+	assert_eq!(data.pop_last(),  Some((0xFF, 0x00)));
+}
+
+#[test]
 fn test_identity_set_serialise_deserialise() {
 	let input = IdentityMap::<char, [u8; 0x2]>::from([
+		('Ä', *b"AE"),
+		('Ö', *b"OE"),
+		('Ü', *b"UE"),
 		('Æ', *b"AE"),
 		('Ø', *b"OE"),
 		('Å', *b"AA"),
+		('ä', *b"ae"),
+		('ö', *b"oe"),
+		('ü', *b"ue"),
 		('æ', *b"ae"),
 		('ø', *b"oe"),
 		('å', *b"aa"),